+++ src/main/java/ru/vyarus/dropwizard/guice/module/lifecycle/event/ConfigurationPhaseEvent.java
@@ -5,18 +5,18 @@ import ru.vyarus.dropwizard.guice.module.context.option.Options;
 import ru.vyarus.dropwizard.guice.module.lifecycle.GuiceyLifecycle;
 
 /**
- * Base class for dropwizard initialization phase events.
+ * Base class for dropwizard configuration phase events.
  *
  * @author Vyacheslav Rusakov
  * @since 13.06.2019
  */
-public abstract class InitPhaseEvent extends GuiceyLifecycleEvent {
+public abstract class ConfigurationPhaseEvent extends GuiceyLifecycleEvent {
 
     private final Bootstrap bootstrap;
 
-    public InitPhaseEvent(final GuiceyLifecycle type,
-                          final Options options,
-                          final Bootstrap bootstrap) {
+    public ConfigurationPhaseEvent(final GuiceyLifecycle type,
+                                   final Options options,
+                                   final Bootstrap bootstrap) {
         super(type, options);
         this.bootstrap = bootstrap;
     }